models refactoring to make them overridable
