+++ setup.py
@@ -13,6 +13,29 @@ version = djangocms_blog.__version__
 readme = open('README.rst').read()
 history = open('HISTORY.rst').read().replace('.. :changelog:', '')
 
+install_requires = [
+    'django-parler>=1.2',
+    'django-cms>3.0.11',
+    'django-taggit>=0.12.2',
+    'django-filer',
+    'pytz',
+    'django-taggit-templatetags',
+    'django-taggit-autosuggest',
+    'djangocms-text-ckeditor',
+    'cmsplugin-filer',
+    'django-meta>=0.2',
+    'django-meta-mixin>=0.2.1',
+    'aldryn-apphooks-config>=0.2.6',
+    'djangocms-apphook-setup',
+    'aldryn-search'
+]
+
+try:
+    import importlib
+    assert importlib
+except ImportError:
+    install_requires.append('importlib')
+
 
 setup(
     name='djangocms-blog',
@@ -26,22 +49,7 @@ setup(
         'djangocms_blog',
     ],
     include_package_data=True,
-    install_requires=[
-        'django-parler>=1.2',
-        'django-cms>3.0.11',
-        'django-taggit>=0.12.2',
-        'django-filer',
-        'pytz',
-        'django-taggit-templatetags',
-        'django-taggit-autosuggest',
-        'djangocms-text-ckeditor',
-        'cmsplugin-filer',
-        'django-meta>=0.2',
-        'django-meta-mixin>=0.2.1',
-        'aldryn-apphooks-config>=0.2.6',
-        'djangocms-apphook-setup',
-        'aldryn-search'
-    ],
+    install_requires=install_requires,
     license='BSD',
     zip_safe=False,
     keywords='djangocms-blog, blog, django, wordpress, multilingual',
